Move DimensionCollectionFactory implementation into Application namespace
